+++ examples/continuous_miniseed_creation.py
@@ -23,7 +23,7 @@ Example usage:
   > python continuous_miniseed_creation.py output.mseed
 
 This file is part of the pymseed package.
-Copyright (c) 2025, EarthScope Data Services
+Copyright (c) 2026, EarthScope Data Services
 """
 
 import argparse