+++ examples/read_numpy.py
@@ -17,7 +17,7 @@ Usage:
 > python read_numpy.py [file1.mseed] [file2.mseed] ...
 
 This file is part of the pymseed package.
-Copyright (c) 2025, EarthScope Data Services
+Copyright (c) 2026, EarthScope Data Services
 """
 
 import argparse